+++ packages/plugins/type-distributor/src/ui/components/columns/bay-type/function-type.svelte
@@ -7,21 +7,21 @@ import LNode from './lnode.svelte'
 
 interface Props {
 	func: FunctionTemplate
-	bayTypeInstanceUuid?: string
+    bayTypeInstanceUuid: string
 }
 
 const { func, bayTypeInstanceUuid }: Props = $props()
 
 let isOpen = $state(false)
 let isDragging = $derived(
-	dndStore.isDraggingItem('functionTemplate', func.uuid, bayTypeInstanceUuid)
+    dndStore.isDraggingItem('functionTemplate', func.uuid, bayTypeInstanceUuid)
 )
 
-const parentUuid = $derived(bayTypeInstanceUuid || func.uuid)
+const parentUuid = $derived(bayTypeInstanceUuid)
 
 let assignedStatuses = $derived(
 	func.lnodes.map((lnode) =>
-		assignedLNodesStore.isAssigned(parentUuid, lnode)
+        assignedLNodesStore.isAssigned(parentUuid, lnode, func.uuid)
 	)
 )
 
@@ -34,7 +34,8 @@ function handleDragStart(event: DragEvent) {
 		type: 'functionTemplate',
 		sourceFunction: func,
 		lNodes: func.lnodes || [],
-		bayTypeInstanceUuid: bayTypeInstanceUuid
+        parentUuid,
+        functionScopeUuid: func.uuid
 	})
 }
 
@@ -47,7 +48,8 @@ function handleLNodeDragStart(event: DragEvent, lnode: LNodeTemplate) {
 		type: 'lNode',
 		sourceFunction: func,
 		lNodes: [lnode],
-		bayTypeInstanceUuid: bayTypeInstanceUuid
+        parentUuid,
+        functionScopeUuid: func.uuid
 	})
 }